requirements.txt
coreapi==2.1.1
dj-database-url==0.4.1
Django==1.10.4
django-csv-importer==0.1.3.5
django-filter==1.0.1
djangorestframework==3.5.3
djangorestframework-jwt==1.11.0
gunicorn==19.6.0
itypes==1.1.0
psycopg2==2.7.1
PyJWT==1.5.2
python-dateutil==2.6.1
pytz==2017.2
requests==2.12.4
six==1.10.0
uritemplate==3.0.0
whitenoise==3.2.2
Pillow==4.2.1
django-rest-swagger==2.1.2
